Examples/DeepStreamVision/detector/requirements.txt
flask>=2.0.0
prometheus-client>=0.16.0
numpy>=1.21.0
Pillow>=9.0.0
requests>=2.28.0
